+++ NEWS.md
@@ -1,6 +1,7 @@
 # `reReg` 1.4.6
 	* Changed maintainer email
 	* Changed last.bumpup to right.polygons in geom_dl() when showName is specified
+	* Scale extreme covariates for enhanced stability
 # `reReg` 1.4.5
   	* Now requires R 4.2.0
 	* Added JSS citation/acknowledgement